[libromdata] WonderSwan: Reduced ROM/SRAM size tables from `unsigned int` to `uint16_t`.
This effectively cuts the size requirements in half.

Code size differences: (64-bit Gentoo Linux, gcc-12.1.0, release build, no LTO)

   text    data     bss     dec     hex filename
  23067     424       0   23491    5bc3 WonderSwan.cpp.o [before]
  23019     424       0   23443    5b93 WonderSwan.cpp.o [after]
    -48       0       0       0     -30 Difference

There were 10 ROM sizes and 6 SRAM sizes, so I would have expected a
32-byte reduction. I guess the extra 16 bytes is from padding.

(cherry picked from commit 5dae3fb86f7053a01efaad7565ab9cb321a39fb5)

+++ src/libromdata/Handheld/WonderSwan.cpp
@@ -477,7 +477,7 @@ int WonderSwan::loadFieldData(void)
 		v_system_bitfield_names, 0, ws_system);
 
 	// ROM size
-	static const unsigned int rom_size_tbl[] = {
+	static const uint16_t rom_size_tbl[] = {
 		128, 256, 512, 1024,
 		2048, 3072, 4096, 6144,
 		8192, 16384,
@@ -493,7 +493,7 @@ int WonderSwan::loadFieldData(void)
 	}
 
 	// Save size and type
-	static const unsigned int sram_size_tbl[] = {
+	static const uint16_t sram_size_tbl[] = {
 		0, 8, 32, 128, 256, 512,
 	};
 	const char *const save_memory_title = C_("WonderSwan", "Save Memory");
